+++ server-adapters/koa/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @mastra/koa
 
+## 1.6.7-alpha.1
+
+### Patch Changes
+
+- Updated dependencies:
+  - @mastra/core@1.50.2-alpha.1
+  - @mastra/server@1.50.2-alpha.1
+
 ## 1.6.7-alpha.0
 
 ### Patch Changes